+++ packages/asset-core-plugin/src/types/plugin/plugin.ts
@@ -16,5 +16,3 @@ export interface IAssetPolishPlugin {
   readonly displayName: string
   polish?: IAssetPluginPolish
 }
-
-export type IAssetPlugin = IAssetParsePlugin & IAssetPolishPlugin